Make more failsafe

+++ src/main/java/cam/lechner/budgetexchange/application/SendBillToCospend.java
@@ -4,7 +4,10 @@ import cam.lechner.budgetexchange.apicall.ApiCall;
 import cam.lechner.budgetexchange.entity.MapCategory;
 import cam.lechner.budgetexchange.entity.TransactionIds;
 import cam.lechner.budgetexchange.entity.Transaktion;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.logging.LogLevel;
 import org.springframework.stereotype.Component;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
@@ -13,6 +16,7 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
 
 @Component
 public class SendBillToCospend {
@@ -28,12 +32,14 @@ public class SendBillToCospend {
     private MapCategoryRepository mapCategoryRepository;
     @Autowired
     private MapMemberRepository mapMemberRepository;
+    private static final Logger LOG = LoggerFactory.getLogger(SendBillToCospend.class);
 
     public void getMissingTransactionsAndSendToCospend() {
         Calendar cal = Calendar.getInstance();
         SimpleDateFormat formater = new SimpleDateFormat("yyyy-MM-dd");
         List<MapCategory> t = new ArrayList<MapCategory>();
         List<TransactionIds> transIds = new ArrayList<TransactionIds>();
+        final Boolean[] errorOccured = {false};
         mapCategoryRepository.findAll().forEach(t::add);
 
         compareRepository.findAll().forEach(transIds::add);
@@ -47,37 +53,20 @@ public class SendBillToCospend {
         //compareRepository.updateIsChecked(0);
 
         t.forEach(kat -> {
-            String projectId = kat.getProjectname();
-            List<Transaktion> trans = apicall.getTransactionWithCategoryAndDate(kat.getBudgetCategory() + "", "2013-01-01", formater.format(cal.getTime()));
-            String payer = getPayer(kat);
-            String payed_for = getPayedFor(kat);
-            for (Transaktion tr : trans) {
-                TransactionIds transactionIds = new TransactionIds();
-                transactionIds = compareRepository.findByBudgetTransId(tr.getId());
-                if (transactionIds == null) {
-                    if (notToCalculate(tr)) {
-                        continue;
-                    }
-                    TransactionIds newtransactionIds = new TransactionIds();
-                    newtransactionIds.setBudgetTransId(tr.getId());
-                    MultiValueMap<String, String> map = new LinkedMultiValueMap<String, String>();
-                    map.add("amount", tr.getWert() * kat.getInout() + "");
-                    map.add("what", tr.getName());
-                    map.add("payer", payer);
-                    map.add("repeat", "n");
-                    map.add("payed_for", payed_for);
-                    map.add("date", tr.getDatum());
-                    // map.add("categoryid", mapCategoryRepository.findByBudgetCategory(tr.getKategorie()).getCospendCategory() + "");
-                    map.add("categoryid", kat.getCospendCategory() + "");
-                    newtransactionIds.setNextcloudBillId(apicall.sendBill(map,projectId));
-                    newtransactionIds.setIsChecked(1);
-                    newtransactionIds.setProjectId(projectId);
-                    compareRepository.save(newtransactionIds);
-                    transaktionRepository.save(tr);
-                } else {
-
-                    Transaktion storedTrans = transaktionRepository.findById(transactionIds.getBudgetTransId()).orElseThrow();
-                    if (!storedTrans.getName().equals(tr.getName()) || storedTrans.getWert() != tr.getWert()) {
+            try {
+                String projectId = kat.getProjectname();
+                List<Transaktion> trans = apicall.getTransactionWithCategoryAndDate(kat.getBudgetCategory() + "", "2013-01-01", formater.format(cal.getTime()));
+                String payer = getPayer(kat);
+                String payed_for = getPayedFor(kat);
+                for (Transaktion tr : trans) {
+                    TransactionIds transactionIds = new TransactionIds();
+                    transactionIds = compareRepository.findByBudgetTransId(tr.getId());
+                    if (transactionIds == null) {
+                        if (notToCalculate(tr)) {
+                            continue;
+                        }
+                        TransactionIds newtransactionIds = new TransactionIds();
+                        newtransactionIds.setBudgetTransId(tr.getId());
                         MultiValueMap<String, String> map = new LinkedMultiValueMap<String, String>();
                         map.add("amount", tr.getWert() * kat.getInout() + "");
                         map.add("what", tr.getName());
@@ -85,28 +74,52 @@ public class SendBillToCospend {
                         map.add("repeat", "n");
                         map.add("payed_for", payed_for);
                         map.add("date", tr.getDatum());
-                        map.add("id", +transactionIds.getNextcloudBillId() + "");
+                        // map.add("categoryid", mapCategoryRepository.findByBudgetCategory(tr.getKategorie()).getCospendCategory() + "");
                         map.add("categoryid", kat.getCospendCategory() + "");
-                        apicall.updateBill(map, "/" + transactionIds.getNextcloudBillId(),projectId);
+                        newtransactionIds.setNextcloudBillId(apicall.sendBill(map, projectId));
+                        newtransactionIds.setIsChecked(1);
+                        newtransactionIds.setProjectId(projectId);
+                        compareRepository.save(newtransactionIds);
                         transaktionRepository.save(tr);
+                    } else {
 
+                        Transaktion storedTrans = transaktionRepository.findById(transactionIds.getBudgetTransId()).orElseThrow();
+                        if (!storedTrans.getName().equals(tr.getName()) || storedTrans.getWert() != tr.getWert()) {
+                            MultiValueMap<String, String> map = new LinkedMultiValueMap<String, String>();
+                            map.add("amount", tr.getWert() * kat.getInout() + "");
+                            map.add("what", tr.getName());
+                            map.add("payer", payer);
+                            map.add("repeat", "n");
+                            map.add("payed_for", payed_for);
+                            map.add("date", tr.getDatum());
+                            map.add("id", +transactionIds.getNextcloudBillId() + "");
+                            map.add("categoryid", kat.getCospendCategory() + "");
+                            apicall.updateBill(map, "/" + transactionIds.getNextcloudBillId(), projectId);
+                            transaktionRepository.save(tr);
+
+                        }
+                        transactionIds.setIsChecked(1);
+                        compareRepository.save(transactionIds);
                     }
-                    transactionIds.setIsChecked(1);
-                    compareRepository.save(transactionIds);
 
                 }
-
+            } catch (Exception e ) {
+                LOG.error(" Exception " +e);
+                apicall.sendMessageToTalk("[Cospend] !!!! Fehler  +e");
+                errorOccured[0] = true;
             }
         });
-        List<TransactionIds> deleteList = compareRepository.findByIsChecked(0);
-        deleteList.forEach(deleteBill -> {
-            apicall.deleteBill(deleteBill.getNextcloudBillId() + "",deleteBill.getProjectId());
-            compareRepository.delete(deleteBill);
-            transaktionRepository.deleteById(deleteBill.getBudgetTransId());
-        });
+        if ( ! errorOccured[0]) {
+            List<TransactionIds> deleteList = compareRepository.findByIsChecked(0);
+            deleteList.forEach(deleteBill -> {
+                apicall.deleteBill(deleteBill.getNextcloudBillId() + "", deleteBill.getProjectId());
+                compareRepository.delete(deleteBill);
+                transaktionRepository.deleteById(deleteBill.getBudgetTransId());
+            });
+        }
     }
 
-    private String getPayer(MapCategory map) {
+    private String getPayer(MapCategory map) throws Exception{
 
         if (map.getKind() == 0) {
             //Miete
@@ -129,8 +142,16 @@ public class SendBillToCospend {
     private Boolean notToCalculate(Transaktion trans) {
         if (trans.getKategorie() != 89 && trans.getKategorie() != 122
             && trans.getKategorie() != 39
-        && trans.getKategorie() != 71) {
+                && trans.getKategorie() != 139
+                && trans.getKategorie() != 142
+                && trans.getKategorie() != 71) {
             return false;
+        }
+        if (trans.getKategorie()== 139  && trans.getKonto_id() != 32) {
+            return true;
+        }
+        if (trans.getKategorie()== 142  && trans.getKonto_id() != 32) {
+            return true;
         }
         if (trans.getKategorie()== 71  && isNotMugRueckzahlung(trans)) {
             return true;
@@ -170,7 +191,7 @@ public class SendBillToCospend {
         return false;
     }
 
-    private String getPayedFor(MapCategory map) {
+    private String getPayedFor(MapCategory map) throws Exception {
 
         if (map.getKind() == 0) {
             //Miete
